Fix boot script for FIT-images
The boot script attempted to load the FIT image too close to the kernel
load address. This causes the kernel image to overwrite FIT image once
it gets extracted. Load the FIT image to initrd address as it is not
used, is sufficiently large and far away from kernel load address in
memory space.